+++ tests/testthat/test_modify_xml.R
@@ -12,6 +12,10 @@ test_that("show_xml works as expected", {
     x3p_show_xml(logo, "no_field_by_this_name"),
     "no fields containing .no_field_by_this_name. found"
   )
+  expect_equal(
+    x3p_show_xml(logo, 1),
+    list(sizeY = 419)
+  )
 })
 
 
@@ -23,14 +27,19 @@ test_that("modify_xml works as expected", {
     x3p_show_xml(logo2, "creator"),
     list(Creator = "someone_else")
   )
-
+  
   expect_error(
     x3p_modify_xml(logo, "x"),
+    'argument "value" is missing, with no default'
+  )
+
+  expect_error(
+    x3p_modify_xml(logo, "x", 1),
     "More than one element matching <x> found"
   )
 
   expect_error(
-    x3p_modify_xml(logo, "nonsense_field"),
+    x3p_modify_xml(logo, "nonsense_field", 1),
     "No element found matching <nonsense_field>"
   )
 })